fix: adjustment in the price input, starting to receive float values instead of just integers

+++ database/migrations/2023_05_30_162249_create_incomes_table.php
@@ -14,7 +14,7 @@ return new class() extends Migration {
             $table->id();
             $table->integer('itemQnt');
             $table->string('itemName');
-            $table->integer('itemPrice');
+            $table->decimal('itemPrice', 10, 2);
             $table->decimal('totalGains', 10, 2)->default(0.00);
             $table->decimal('totalExpenses', 10, 2)->default(0.00);
             $table->timestamps();
